Reorganize impl tests
